Language configuration per module.
Adds a column module_lang to the table module to be able to
define gettext language per module. A default, which is used when no
module can be found, can be set with APP__DEFAULT_LOCALE in
inc_global.php. The language can be changed for a module in the admin
section for modules (which will list all folders of locale/ as possible
languages). (It is only a quick implementation, it does not do any
extensive checks.)

The webpa2_administrator.sql file has also been modified to add a
row into the user_module table because without this row I could not
login into a new webPA installation.